Support pagination with result grouping
When using [result grouping](https://cwiki.apache.org/confluence/display/solr/Result+Grouping), we don't get a simple number of documents matched. Instead, related documents are collapsed into groups and the groups are what we (probably) want to paginate.

This is a bit whacky as it is technically possible to specify different grouping criteria. In that case, we'd get multiple lists of groups back. 

This change only looks at the first grouping applied as I don't see how we could make sense of the multi-group case.